BUG: Allow env without python-tk.
When running in a mode which does not generate graphs, the modules supporting
plotting are not needed for import.

Catch and warn when a module required for plotting can not be imported.

This fixes the following exception during import of the baysealpha.
```
  File "/usr/lib/python2.7/lib-tk/Tkinter.py", line 42, in <module>
    raise ImportError, str(msg) + ', please install the python-tk package'
ImportError: No module named _tkinter, please install the python-tk package
```